+++ tests/test_letter_timings.py
@@ -5,7 +5,15 @@ import pytz
 from freezegun import freeze_time
 
 from notifications_utils.letter_timings import (
+    get_dvla_working_day_offset_by,
     get_letter_timings,
+    get_next_dvla_working_day,
+    get_next_royal_mail_working_day,
+    get_previous_dvla_working_day,
+    get_previous_royal_mail_working_day,
+    get_royal_mail_working_day_offset_by,
+    is_dvla_working_day,
+    is_royal_mail_working_day,
     letter_can_be_cancelled,
 )
 
@@ -364,3 +372,59 @@ def test_letter_can_be_cancelled_if_after_1730_and_letter_created_at_1730_today_
 )
 def test_letter_can_be_cancelled_always_compares_in_bst(notification_created_at):
     assert letter_can_be_cancelled("created", notification_created_at)
+
+
+def test_next_previous_working_days():
+    friday_23_december = datetime(2022, 12, 23, 12, 0, 0)
+    saturday_24_december = datetime(2022, 12, 24, 12, 0, 0)
+    sunday_25_december = datetime(2022, 12, 25, 12, 0, 0)
+    monday_26_december = datetime(2022, 12, 26, 12, 0, 0)
+    tuesday_27_december = datetime(2022, 12, 27, 12, 0, 0)
+    wednesday_28_december = datetime(2022, 12, 28, 12, 0, 0)
+
+    # DVLA and Royal Mail don’t work on Sundays or bank holidays
+    assert not is_dvla_working_day(sunday_25_december)
+    assert not is_dvla_working_day(monday_26_december)
+    assert not is_dvla_working_day(tuesday_27_december)
+    assert get_next_dvla_working_day(monday_26_december) == wednesday_28_december
+
+    assert not is_royal_mail_working_day(sunday_25_december)
+    assert not is_royal_mail_working_day(monday_26_december)
+    assert not is_royal_mail_working_day(tuesday_27_december)
+    assert get_next_royal_mail_working_day(monday_26_december) == wednesday_28_december
+
+    # DVLA don’t work Saturdays
+    assert not is_dvla_working_day(saturday_24_december)
+    assert get_previous_dvla_working_day(monday_26_december) == friday_23_december
+
+    # Royal Mail do work Saturdays
+    assert is_royal_mail_working_day(saturday_24_december)
+    assert get_previous_royal_mail_working_day(monday_26_december) == saturday_24_december
+
+
+def test_get_offset_working_day():
+    friday_16_december = datetime(2022, 12, 16, 12, 0, 0)
+    saturday_17_december = datetime(2022, 12, 17, 12, 0, 0)
+    thursday_22_december = datetime(2022, 12, 22, 12, 0, 0)
+    friday_23_december = datetime(2022, 12, 23, 12, 0, 0)
+    wednesday_28_december = datetime(2022, 12, 28, 12, 0, 0)
+    thursday_29_december = datetime(2022, 12, 29, 12, 0, 0)
+
+    # 1 day forward, skipping weekend and bank holidays
+    assert get_dvla_working_day_offset_by(friday_23_december, days=1) == wednesday_28_december
+    # 2 day forward, skipping weekend, bank holidays, and 1 working day
+    assert get_dvla_working_day_offset_by(friday_23_december, days=2) == thursday_29_december
+
+    # 1 day backward, skipping no days
+    assert get_dvla_working_day_offset_by(friday_23_december, days=-1) == thursday_22_december
+    # 5 days backward, skipping weekend
+    assert get_dvla_working_day_offset_by(friday_23_december, days=-5) == friday_16_december
+    # 5 days backward, skipping Saturday only
+    assert get_royal_mail_working_day_offset_by(friday_23_december, days=-5) == saturday_17_december
+
+
+@pytest.mark.parametrize("function", (get_dvla_working_day_offset_by, get_royal_mail_working_day_offset_by))
+def test_get_0_offset_working_days(function):
+    with pytest.raises(ValueError) as error:
+        function(datetime.now(), days=0)
+    assert str(error.value) == "days argument must not be 0"